+++ packages/create/src/generators/wc-lit-element/index.js
@@ -1,9 +1,34 @@
+import { join } from 'path';
 import { CommonRepoMixin } from '../common-repo/index.js';
+import { processTemplate, readFileFromPath } from '../../core.js';
+
+const compose = (...fns) => fns.reduce((f, g) => (...args) => f(g(...args)));
+const safeReduce = (f, initial) => xs => (Array.isArray(xs) ? xs.reduce(f, initial) : xs);
+
+const getTemplatePart = compose(
+  processTemplate,
+  readFileFromPath,
+);
+
+function featureReadmeBlurb(feature) {
+  const path = join(__dirname, `./templates/partials/README.${feature}.md`);
+  return getTemplatePart(path);
+}
+
+function featureReadme(acc, feature, i, a) {
+  return `${acc + featureReadmeBlurb(feature)}${i === a.length - 1 ? '' : '\n'}`;
+}
+
+const safeFeatureReadme = safeReduce(featureReadme, '');
 
 /* eslint-disable no-console */
 export const WcLitElementMixin = subclass =>
   class extends subclass {
     async execute() {
+      this.templateData.featureReadmes = safeFeatureReadme(this.options.features);
+      this.templateData.scriptRunCommand =
+        this.options.installDependencies === 'yarn' ? 'yarn' : 'npm run';
+
       await super.execute();
       const { tagName, className } = this.templateData;