+++ packages/semantic-dom-diff/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.3](https://github.com/open-wc/open-wc/compare/@open-wc/semantic-dom-diff@0.15.2...@open-wc/semantic-dom-diff@0.15.3) (2019-11-02)
+
+**Note:** Version bump only for package @open-wc/semantic-dom-diff
+
+
+
+
+
 ## [0.15.2](https://github.com/open-wc/open-wc/compare/@open-wc/semantic-dom-diff@0.15.1...@open-wc/semantic-dom-diff@0.15.2) (2019-10-25)